Dang you Rednaxela! Dang you to heck! :-)
Let's try this commit again after having to update the entire map tree thanks to Rednaxela.

Update to Zorn Castle floor two. New modified monsters and some generators added. Also mods to the floor plan. The exits to the towers and to the top floor do not yet work. This floor is basically "finished" and playable.


git-svn-id: svn://svn.code.sf.net/p/crossfire/code/maps@5185 282e977c-c81d-0410-88c4-b93c2d0d6712