Describe the origin or purpose of every key
This patch will give the vast majority of the keys in the game
a message that describes the location the key was found in, or straight-up what it's used for.

It also names or renames a lot of them, although this isn't the focus (for now).

Patch https://sourceforge.net/p/crossfire/patches/415/ by Saiapatsu.