Change save beds to normal beds
Set beds to "normal" beds rather than save beds here in Goth's Tavern as
some of them are above other map items and when a player would press "a"
to active the item, they would instead get logged out because of the
bed. If a player wants a Save Point in Scorn, using Wanderer's Inn is
probably adequate, and a single location for save/restore points in
Scorn is much easier to direct new players to.