Close old private shops
These private shops have been broken for some time, so close them in
preparation for something new to arrive.

Because these are unique maps and because players might still have stuff
on them, we can't "just" delete them. So close them off and add a sign
to contact a dungeon master if they need anything recovered.

Eventually these will be unlinked but will have to remain at the same
path in case items need to be recovered in the future.